Make `Perm#apply(k: Int): Int` the left action
Function application syntax, `p(n)` is the left action. Since this is
the default action of Perm on Int, `combine` should be implemented in
terms of `compose` rather than `andThen`.

Given an intetger `n`, the left group action is denoted `p·n`, so that
`(pr)⋅x == p·(r·n)` for all permutations `p` and `r`.

The difference between left and right actions is in the order in which
a product like gh acts on x. For a left action h acts first and is
followed by g, while for a right action g acts first and is followed
by h.

`Perm#permute` is also corrected so that

```
Perm(0, 2, 3).permute((1 to 5).toSeq) == Seq(4, 2, 0, 3, 5)
```

The cycle constructor now requires `cycles` that are disjoint.  Choice
of `compose` or `andThen` no longer matters, since multiplication of
disjoint cycles is commutative.